+++ packages/commerce-sdk-react/src/hooks/useCustomerId.ts
@@ -7,7 +7,8 @@
 import useAuthContext from './useAuthContext'
 import useLocalStorage from './useLocalStorage'
 import useConfig from './useConfig'
-import {onClient} from '../utils'
+
+const onClient = typeof window !== 'undefined'
 
 /**
  * Hook that returns the customer ID.
@@ -16,7 +17,7 @@ import {onClient} from '../utils'
  * @category Shopper Authentication
  */
 const useCustomerId = (): string | null => {
-    if (onClient()) {
+    if (onClient) {
         // `onClient` is a constant, so the hooks will always have the same execution order,
         // despite technically being inside a conditional.
         // eslint-disable-next-line react-hooks/rules-of-hooks