+++ packages/commerce-sdk-react/src/hooks/useEncUserId.ts
@@ -7,7 +7,8 @@
 import useAuthContext from './useAuthContext'
 import useLocalStorage from './useLocalStorage'
 import useConfig from './useConfig'
-import {onClient} from '../utils'
+
+const onClient = typeof window !== 'undefined'
 
 /**
  * @group Shopper Authentication helpers
@@ -30,7 +31,7 @@ const useEncUserId = (): EncUserId => {
     const config = useConfig()
     const auth = useAuthContext()
 
-    const encUserId = onClient()
+    const encUserId = onClient
         ? // This conditional is a constant value based on the environment, so the same path will
           // always be followed., and the "rule of hooks" is not violated.
           // eslint-disable-next-line react-hooks/rules-of-hooks