+++ packages/commerce-sdk-react/src/provider.tsx
@@ -46,6 +46,7 @@ export interface CommerceApiProviderProps extends ApiClientConfigParams {
     passwordlessLoginCallbackURI?: string
     refreshTokenRegisteredCookieTTL?: number
     refreshTokenGuestCookieTTL?: number
+    OCAPISessionsURL?: string
 }
 
 /**