+++ packages/commerce-sdk-react/CHANGELOG.md
@@ -1,3 +1,8 @@
+## v3.3.0-extensibility-preview.5 (May 06, 2025)
+
+## v3.3.0-extensibility-preview.4 (Feb 12, 2025)
+- Add `ServerContext` type for `useServerContext` hook [#2239](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2239)
+
 ## v3.3.0-dev.1 (Apr 30, 2025)
 - Fix inconsistency between dwsid and access token for guest login when hybrid authentication is enabled [#2397](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2397)